requirements.txt
langchain
langchain_community
openai
urllib3==1.26.17
weaviate-client
google-generativeai>=0.8.0
pillow